pipeline/data_extract_load/file_loader/requirements.txt
google-cloud-bigquery
google-cloud-storage
functions-framework